requirements.txt
# ── Core multi-agent framework ─────────────────────────────────────────────────
crewai==0.177.0

# ── LLM providers ──────────────────────────────────────────────────────────────
# CrewAI routes through LiteLLM internally; anthropic is explicitly listed for
# version clarity with the claude-sonnet-4-6 model.
anthropic>=0.84.0,<1.0.0
litellm>=1.50.0,<2.0.0

# ── Environment management ─────────────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0.0

# ── Validation (CrewAI requires pydantic v2) ───────────────────────────────────
pydantic>=2.10.0,<3.0.0

# ── Trading: Alpaca SDK ────────────────────────────────────────────────────────
# alpaca-py is the official, actively maintained SDK (alpaca-trade-api is deprecated)
alpaca-py>=0.38.0,<1.0.0

# ── Trading: Telegram Bot ──────────────────────────────────────────────────────
python-telegram-bot>=22.7,<23.0

# ── Trading: Data processing & technical analysis ──────────────────────────────
pandas>=2.0.0,<3.0.0
ta>=0.11.0,<1.0.0

# ── Trading: Async HTTP & database ────────────────────────────────────────────
aiohttp>=3.9.0,<4.0.0
aiosqlite>=0.20.0,<1.0.0
